fix(config): support fractional dependency timeout instead of truncating to 0
`toSeconds` built a time.Duration as `time.Duration(val) * time.Second`,
which truncates a float64 to a whole second before scaling. Any value in
(0,1) collapsed to 0 — and 0 means "wait indefinitely" — so
`depends-on "a" timeout=0.5` requested a half-second bound and silently
got an unbounded wait, with no error/warning/log. That is a Silent
Failure Prohibition violation in a function whose own comments invoke the
rule: a truncation that turns a bound into its absence.

The downstream consumer (daemon_autostart.go) already feeds dep.Timeout
straight into context.WithTimeout, so sub-second support is end-to-end and
non-invasive — the honest fix that matches user intent rather than
refusing the value. Multiply by float64(time.Second) before the Duration
conversion to keep the fraction: 0.5 -> 500ms, 1.5 -> 1500ms.

The 0 sentinel is deliberately preserved: an explicit timeout=0 (and an
omitted timeout) still means wait indefinitely; only the
fractional-truncation-to-0 case changes. docs/configuration.md documents
the sub-second granularity and the 0 sentinel.

worktrack-task: 01KYSZ89VXGFQD9GNG34WK1KB4

+++ internal/config/agnt_deps.go
@@ -103,6 +103,17 @@ func (d *DependsOnList) UnmarshalKDL(node *document.Node) error {
 }
 
 // toSeconds converts a numeric value to a time.Duration in seconds.
+//
+// Fractional values are preserved with sub-second granularity: `timeout=0.5`
+// yields 500ms, not 0. Multiplying a float64 by float64(time.Second) before the
+// Duration conversion is what keeps the fraction — the previous
+// `time.Duration(val) * time.Second` truncated the float to a whole second
+// first, so any value in (0,1) collapsed to the 0 == "wait indefinitely"
+// sentinel. That silent truncation turned a bound into its absence, a Silent
+// Failure Prohibition violation (.claude/rules/daemon-architecture.md).
+//
+// The 0 sentinel is deliberately preserved: an explicit `timeout=0` still means
+// "wait indefinitely". Only the fractional-truncation-to-0 case changes.
 func toSeconds(v interface{}) (time.Duration, bool) {
 	switch val := v.(type) {
 	case int:
@@ -110,7 +121,7 @@ func toSeconds(v interface{}) (time.Duration, bool) {
 	case int64:
 		return time.Duration(val) * time.Second, true
 	case float64:
-		return time.Duration(val) * time.Second, true
+		return time.Duration(val * float64(time.Second)), true
 	}
 	return 0, false
 }
